Refactor and improve code structure
- Moved `swap` function from `index.ts` to `actions.ts`

- Relocated `getWethAddressByChain()` and `getOrderSignerAddressChain()` to `utils.ts`

- Added return types to `getWethAddressByChain()` and `getOrderSignerAddressByChain()`
